requirements.txt
requests==2.31.0
ics==0.7.2
python-dateutil==2.9.0.post0
